LayoutTests/media/video-main-content-autoplay-expected.txt

RUN(smallVideo.setAttribute('controls','controls'))
RUN(smallVideo.setAttribute('autoplay','autoplay'))
RUN(smallVideo.setAttribute('id','small'))
RUN(document.body.appendChild(smallVideo))

RUN(largeVideo.setAttribute('controls','controls'))
RUN(largeVideo.setAttribute('autoplay','autoplay'))
RUN(largeVideo.setAttribute('id','large'))
RUN(document.body.appendChild(largeVideo))

Large video auto-played, small video did not. OK
END OF TEST
